Separated search into separate file (#117)
* Separated Search from Home

* Merged with master
